# Use 64 bit hash and fix out of bounds access