There don't seem to be any statistically significant changes in performance. I filtered Ruby processes and ran `loop { puts :a }` in a loop (see test plan).

Test Plan
=======

https://pprof.me/fc0fff3/

```
sudo bpftool prog | grep profile_cpu
```

```
2938: perf_event  name profile_cpu  tag 4554576c1d62dc61  gpl run_time_ns 136221451 run_cnt 24408 -> 5581.02 ns / run
              2268 run_cnt
          60823512 cycles        -> 26818.13 cycles / run
          42618949 instructions    -> 18791.42 instructions / run #     0.70 insns per cycle
```

```
3004: perf_event  name profile_cpu  tag 75f4ac73b948f3c0q  gpl run_time_ns 144678180 run_cnt 23700 -> 6104.56 ns / run

             2279 run_cnt
          60910573 cycles            -> 26726.89 cycles / run
          42422810 instructions       -> 18614.66 instructions / run #     0.70 insns per cycle
```